+++ utils/TimeCountUtil.go
@@ -32,9 +32,9 @@ func DurationToString(duration time.Duration) string {
 }
 
 //调用次方法 必须加上 defer 关键字！
-func CountMethodTps(total float64, now time.Time, info string) {
+func CountMethodTps(total int, now time.Time, info string) {
 	var end = time.Now()
-	fmt.Println(info+` tps =`, total/end.Sub(now).Seconds())
+	fmt.Println(info+` tps =`, float64(total)/end.Sub(now).Seconds())
 }
 
 func PrintTimeString(info string, start time.Time, end time.Time, duration time.Duration) string {
